+++ service/controller/app/app.go
@@ -18,7 +18,6 @@ import (
 	"github.com/giantswarm/app-operator/v5/pkg/project"
 	"github.com/giantswarm/app-operator/v5/service/controller/app/controllercontext"
 	"github.com/giantswarm/app-operator/v5/service/internal/clientcache"
-	"github.com/giantswarm/app-operator/v5/service/internal/crdcache"
 )
 
 const appControllerSuffix = "-app"
@@ -27,7 +26,6 @@ type Config struct {
 	Fs          afero.Fs
 	K8sClient   k8sclient.Interface
 	ClientCache *clientcache.Resource
-	CRDCache    *crdcache.Resource
 	Logger      micrologger.Logger
 
 	ChartNamespace    string
@@ -50,9 +48,6 @@ func NewApp(config Config) (*App, error) {
 
 	if config.ClientCache == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.ClientCache must not be empty", config)
-	}
-	if config.CRDCache == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CRDCache must not be empty", config)
 	}
 	if config.Fs == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Fs must not be empty", config)
@@ -106,7 +101,6 @@ func NewApp(config Config) (*App, error) {
 	{
 		c := appResourcesConfig{
 			ClientCache: config.ClientCache,
-			CRDCache:    config.CRDCache,
 			FileSystem:  config.Fs,
 			K8sClient:   config.K8sClient,
 			Logger:      config.Logger,