+++ service/controller/app/resource/chartcrd/resource.go
@@ -3,8 +3,6 @@ package chartcrd
 import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-
-	"github.com/giantswarm/app-operator/v5/service/internal/crdcache"
 )
 
 const (
@@ -12,27 +10,21 @@ const (
 )
 
 type Config struct {
-	CRDCache *crdcache.Resource
-	Logger   micrologger.Logger
+	Logger micrologger.Logger
 }
 
 type Resource struct {
-	crdCache *crdcache.Resource
-	logger   micrologger.Logger
+	logger micrologger.Logger
 }
 
 // New creates a new configured tcnamespace resource.
 func New(config Config) (*Resource, error) {
-	if config.CRDCache == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CRDCache must not be empty", config)
-	}
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
 	r := &Resource{
-		crdCache: config.CRDCache,
-		logger:   config.Logger,
+		logger: config.Logger,
 	}
 
 	return r, nil