+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [5.4.1] - 2022-01-14
+
+### Fixed
+
+- Embed Chart CRD in app-operator to prevent hitting GitHub API rate limits.
+
 ## [5.4.0] - 2021-12-17
 
 ### Changed
@@ -418,7 +424,8 @@ from Helm 2 to Helm 3.
 
 - Flattening operator release structure.
 
-[Unreleased]: https://github.com/giantswarm/app-operator/compare/v5.4.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-operator/compare/v5.4.1...HEAD
+[5.4.1]: https://github.com/giantswarm/app-operator/compare/v5.4.0...v5.4.1
 [5.4.0]: https://github.com/giantswarm/app-operator/compare/v5.3.1...v5.4.0
 [5.3.1]: https://github.com/giantswarm/app-operator/compare/v5.3.0...v5.3.1
 [5.3.0]: https://github.com/giantswarm/app-operator/compare/v5.2.0...v5.3.0